fixed bug by implementing time counter
if there are nest loops in user code, we want the context to be natural.
e.g.)
for (var i = 0; i < 3; i++) {
    for (var j = 0; j < 3 j++) {
    }
}
In this case, we assume that the context of out side loop is 3.
The context of inner side loop should be 7 8 or 9.
But, the context of inner side loop might others(e.g. 3)
At this time, the context of inner side loop is changed naturally.