modified how to name the object's ID.
Before, I had given the object's ID NewExpression's ID and the call
times. But, previous implementation don't save context. In the case that a method call is inserted many method calls,
the called count of NewExpression in object's ID was incremented.
So, We added the ID of the called method and the called count of the
method to object's ID.