modified the implementation by using not indexOf but hash.
This commit has some modification.
First, I modified the implementation by using not indexOf but hash so that the evaluation time with checkpoints in Kanon becomes more fast.
Second, I enable users to write some infinite loop. If a user write a infinite loop (like while(true){}), Kanon evaluates the loop 1000 times and stops the evaluation. So Kanon visualizes the object graph during the execution of the code even if a user writes infinite loop.
Third, if you write infinite loop, the text('Infinite loop?') is displayed in the context view of the loop.
Finally, I fixed some bugs.