do not discard implementation errors
update.js try-catches the 'eval' of user code so as to handle infinite
loops.  It is fine when there is no errors in implementation.
However, when there is, the catch clause discards exceptions other
than infinite loops, which makes it difficult to recognize
implementation errors.

With this patch, it re-throws exceptions other than infinite loops.